Assorted minor changes to modules and readmes that resulted from my work on Dimensional distribution.
Moved some code in BlockArr follower so it is executed less often.
Changed somedomain.dim(1).idxType to somedomain.idxType
(now that idxType is available on domains).

Touched up the comments in ReplicatedDist (thanks Sung for feedback)
and test/distributions/robust/arithmetic/README.

In ReplicatedDist.dsiPrivatize(), ensured that the domain
of the privatized copy of targetLocales is local, too.

  TODO: another way of ensuring that is like it's done in Block:
declare the type of targetLocales to be a default-mapped domain,
and parameterize the distribution descriptor by that domain's
attributes; then create a constructor that provides defaults
for those attribute based on the targetLocales argument.

Added comments on the _value and _valueType fields of _distribution,
_domain, _array, for future generations of compiler hackers.

Replaced _stride -> stride in a couple of places in ChapelRangeBase.chpl.

Edited README.dsi:
* Do not require dsiSupportsPrivatization() for classes that do not
  support privatization - it is inherited from the superclass.
* Introduce pidNotYetPrivatized (formerly -1).
* Minor info additions, terminology edits.

Fixed a typo in README.xt-cle.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18941 3a8e244f-b0f2-452b-bcba-4c88e055c3ca