Commiting latest version of Qthreads (r2201) for test and evaluation.
This version includes a bug fix in chpl_task_callMain() which was
resulting in unnecessary synchronization overhead.

I have not made this the default yet, since I'm not sure what policy
we want to employ.  In the meantime, if you want to use it you'll need
to change QTHREAD_VERSION to r2201 in $CHPL_HOME/make/Makefile.base.
I also didn't check in the configure script, so the first time you use
this version, you have to go into $CHPL_HOME/third-party/qthread and
do 'make qthread-reconfig'.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18943 3a8e244f-b0f2-452b-bcba-4c88e055c3ca