Added ceil(n,m) to the Math module.
Also added a test to verify proper operation (by comparing against
ceil(real)) and an entry in the spec (for the Math module).

verify-ceil-2.chpl contains a couple of combinations of argument types
for which ceil(n,m) presently does not typecheck, but it should.
So it's a feature-request future.

Curiously, part of the problem is our inability to add uint(32) + uint(64).
There are probably similar issues with subtraction and division.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18949 3a8e244f-b0f2-452b-bcba-4c88e055c3ca