Added 'cleandeps' target to runtime/Makefile.
This target is needed to remove the make-depend files 'Makefile.svn.depend'
in various runtime subdirectories. They are currently not removed by
'make clean' and 'make clobber', although arguably they should be.

This target redirects to 'cleandeps' in
runtime/make/Makefile.runtime.commonrules.
It seems like it was an omission from the
"clean, cleandeps, clobber" lineup.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18966 3a8e244f-b0f2-452b-bcba-4c88e055c3ca