Simplify running Chapel programs built with nanox tasking.
When linking a Chapel program that is built with nanox tasking, give the
linker a -rpath option to have it include in the executable the directory
containing the libnanox-chpl.so shared object.  With this done, the user does
not have to specify that directory in the LD_LIBRARY_PATH environment variable
at execution time.  Update the documentation to remove the part that describes
setting LD_LIBRARY_PATH for nanox tasking.

Note that this change assumes that the linker is a gnu-based one.  I believe
that this is a valid assumption on all platforms on which we support nanox
tasking, but if that is not the case this change may have to be reverted or
modified/augmented.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18967 3a8e244f-b0f2-452b-bcba-4c88e055c3ca