Simplify running Chapel programs built with qthreads tasking.
(This is the qthreads analog to r18967 for nanox tasking.)

When linking a Chapel program that is built with qthreads tasking, give the
linker a -rpath option to have it include in the executable the directory
containing the libqthread.so shared object.  With this done, the user does not
have to specify that directory in the LD_LIBRARY_PATH environment variable at
execution time.  Update the documentation to remove the part that describes
setting LD_LIBRARY_PATH for qthreads tasking.

Note that this change assumes that the linker is a gnu-based one.  I believe
that this is a valid assumption on all platforms on which we support qthreads
tasking, but if that is not the case this change may have to be reverted or
modified/augmented.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18968 3a8e244f-b0f2-452b-bcba-4c88e055c3ca