Add initial support for tcmalloc.
This adds version 1.7 of google-perftools, containing the tcmalloc memory
allocator which will be a supported runtime memory layer implementation soon.
My next commit will delete the tcmalloc/google-perftools-1.7.tar.gz tarball
containing the distribution, which is included here just so we have a record
of where the sources originally came from.  The commit after that will add the
actual memory layer to the runtime.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18976 3a8e244f-b0f2-452b-bcba-4c88e055c3ca