Parametrized the size of the workloads to make them dependent on config variables.
Added the corresponding .perfcompopts, .perfexecopts, .perfkeys and
.graph files for the performance run.

- By default, in the correctness run each workload executes: fine (n=100 it.,
delay=1 microsec.); coarse (n=10 it., delay=10000 microsec.); tri (triangular
loop with n=10, delay=(n-i)*100 microsec.); and ran (random workload per
iteration with n=10 and delay=100000*rand() microsec.).

- In the performance run (see config variables in the corresponding
.perfexecopts files) each workload executes: fine (n=1e6 it., delay=1
microsec.); coarse (n=100 it., delay=10000 microsec.); tri (triangular loop
with n=1000, delay=(n-i)*100 microsec.); and ran (random workload per
iteration with n=1000 and delay=100000*rand() microsec.). These are the final
workloads of the PGAS'11 paper, although remember that the array references of
each loop are commented for the performance results presented in that paper.

For the adaptative Work-Stealing iterators (distAdaptativeWS,
distAdaptativeWSv1 and distAdaptativeWSv2) only the times for the Binary
splitting strategy (both on the local range and on the victim range) are
represented in the graphs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18991 3a8e244f-b0f2-452b-bcba-4c88e055c3ca