Adding nightly testing scripts
This snapshots the nightly testing scripts that we've been using on
chpbld01, lnx-hilde, and lnx-caseyb repsectively.  These have been in
chapelu's util/cron directory for awhile but never committed.  The
chpbld01 should improve the XT testing tonight.  I had to manually
source /etc/profile for some reason that I'm still looking for a
rationale for...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18998 3a8e244f-b0f2-452b-bcba-4c88e055c3ca