Added an optional module UtilMath, presently with divceilpos and divfloorpos.
Discussion:

* These two functions do the same as divceil and divfloor
but do not perform run-time checks for positive-ness for better performance.
This makes it the user's responsibility to ensure that both arguments
are positive and not 0.

* TODO: add assertions for the above. I have not added assertions
because at the moment there is no way to turn them off.

* divceilpos() uses an overflow-proof formula
that computes correct result for any signed integers > 0, afaik.
However, the price for this is that it does not handle the case m==0 correctly.

* divfloorpos() is there just for symmetry. (It just runs m/n.)
This one does handle the case m==0 correctly, but I prefer not to allow
that to make it simpler to the user.

* Documentation went into the same place in the Chapel spec
as the Random, Sort and other modules, for consistency.
(None of these modules are part of the language definition proper,
but we just don't have a better place for their specification.)

* Other "Util" modules are welcome.

* Added a test to fix the behavior.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19001 3a8e244f-b0f2-452b-bcba-4c88e055c3ca