Changed the default floating-point I/O to print out e.g. 1.0 for an integer.
created qio_channel_scan_literal and qio_channel_print_literal
for doing io of e.g. "(" or "true"; strings that should not
be put in quotes even if we're quoting.

Added a default implementation of preadv and pwritev
for systems that don't have GLIBC 2.10 or later;
I believe that glibc itself will have the implementation
if it is present. This implementation will also trigger
on e.g. Mac OS X.

runtime/src/common-chpl/Makefile:
Hacky bit of perl to fix the problems I was seeing in compiling --runtime
part (too many statics, no chpl_localeID). Hopefully Tom will be cleaning
up this part later and then the perl can be removed.

Added ioLiteral; this and ioNewline allow easy reading/writing
of literals (e.g. the "(" in a tuple); it's a format error
if we're trying to read and we don't match the string.

Default writeThis is once again constructed in buildDefaultFunctions,
but now it just calls writeThisDefaultImpl, which uses the fielderator
to write the fields. It would be nice if we could create equivalent
functionality entirely in the library, but I was having trouble
getting it to work...

My tests (test/io/ferguson) are passing single-locale
and many tests in test/trivial are passing (304 pass, 42 fail).

Known bugs that are entirely my fault:
 - writeThis doesn't do anything intelligent with superclasses
 - readThis is not currently supported -- pending decisions about
   if it should be channel.readThis(object) or object.readThis(channel),
   and if the format should include class names.
 - I've got stack-traces printing which causes many tests to fail
   because they relied on no stack traces. Also, not all halt or
   assert arguments are printing out for some reason.
 - complex numbers don't get a space put around the '+' or '-'
   which is different from how it used to be.
 - readln(string) will read until end-of-file

Known bugs that I'm waiting for fixes from others:
 - bugs with e.g. _extern get_string(inout ret_string:string) are
   causing compilation failure in multilocale. I'd expect these
   to work if with void get_string(char** ret_string); I expect
   the Chapel compiler to generate code to coerce a multilocale
   string into the char** and back again, since it does that
   already with non-inout arguments or with return arguments.
   There's a test case in test/users/ferguson/extern_string_test.*
   We had some discussion about it, but I think that somebody
   familiar with _extern stuff just needs to have a look.
   
 - channels and files should be reference counted and 
   closed/flushed/freed when they are no longer used. The
   reference counting is not currently working correctly,
   so one has to flush or close manually. I'm expecting that
   Seng-Eun is well on the way to making reference counting
   work right for other things, and that I'll be able
   to fix once that work is complete.




git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/ferguson@19005 3a8e244f-b0f2-452b-bcba-4c88e055c3ca