Fixed bug with passing strings to extern functions for --no-local compiles
This fixes a bug submitted by Michael Ferguson related to passing
strings to external functions.  As with most things regarding strings,
the approach is somewhat unsatisfactory and special-casey.
Essentially, we do some extra work to assert that the string is local
in parallel.cpp (much as we do for extern classes being passed to
externs) and then to unwrap the char* field from the wide string type
at codegen time (expr.cpp).

As part of this change, I introduced a global Type* symbol to cache
the wide string type for easy comparison because it's somewhat of a
special beast (e.g., has an extra field compared to all other wide
classes).

When strings are redone properly, I expect much of this special-casing
will be able to go away.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19010 3a8e244f-b0f2-452b-bcba-4c88e055c3ca