Clarified the 'can not capture overloaded functions as values' compiler error message.
BTW it would be great to report the error 'can not capture methods' here as well,
but it seems like FLAG_METHOD is not yet set at this point.
So for now we get the 'illegal access of first class function' error instead.
(I did not see how to tell, where that error is reported, whether the first
class function in question is a method. Otherwise we could produce
the 'can not capture methods' at this point instead.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19017 3a8e244f-b0f2-452b-bcba-4c88e055c3ca