Making Kristi's Graph 500 code deterministic
In the graph generation step of Kristi's graph500 code, there is a
pair of forall loops that are used to randomly permute some aspects of
the graph when constructing it.  Because these permutations were being
computed randomly, they constituted a race condition that would result
in different input graphs being generated.  For example, if you and I
are both computing permutations, and I decide to permute elements 1
and 2 and you decide to permute elements 2 and 3, the output will
depend on which of us gets to element 2 first.  Given input:

1 : A
2 : B
3 : C

if I get there first, we'll get:

1 : B
2 : C
3 : A

If you get there first, we'll get:

1: C
2: A
3: B

I put a serial statement around these two loops controlled by a new
config param named SHUFFLE_DETERMINISTICALLY to control whether they
should be run serially or in parallel.  It's set to 'true' by default,
but 'false' in our correctness tests.  This should make the testing
stable.

Kristi, please check that what I've done seems reasonable.






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19021 3a8e244f-b0f2-452b-bcba-4c88e055c3ca