Committing a bug with the serial statement
From the .future:

This test demonstrates that 'serial' is not currently serializing
forall loops that iterate over distributed domains/arrays.  My
hypothesis is that this is because of the optimization we have to
handle coforall + on loops with lower overhead.  We're probably
ignoring the serial state in this case...  Investigating now.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19024 3a8e244f-b0f2-452b-bcba-4c88e055c3ca