Fixing bug with serial statement and distributed forall loops
This commit fixes a bug in which distributed forall loops were
ignoring the state of serial statements that squashed parallelism due
to our coforall+on optimization to use non-blocking forks.  My fix was
essentially to generate such calls as a macro rather than a direct
call to the runtime, as we do for other communication routines, and to
have the macro check the serial state before deciding whether to call
the blocking or non-blocking routine.  This seems to fix the issue
(and makes Kristi's graph 500 code generate reproducible graphs when
run on multiple locales now).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19026 3a8e244f-b0f2-452b-bcba-4c88e055c3ca