The module AdvancedIters.chpl contains the dynamic, guided and adaptive work-stealing iterators, supporting non-strided and strided ranges as well as parallel zippered iterations.
- dynamic(range, chunkSize, nTasks) represents the equivalent to OpenMP
dynamic scheduling approach: given an input range, each task (of nTasks)
splits chunks of fixed size (chunkSize) when it finishes its previous chunk.

- guided(range, nTasks) represents the OpenMP guided policy: given an input
range, each task splits chunks of variable size when it finishes its previous
chunk. The size of each chunk is the number of unassigned iterations divided
by the number of tasks (nTasks). These sizes decrease approximately
exponentially to 1. The splitting strategy is therefore adaptive.

- adaptive(range, nTasks) represents a naive adative work-stealing leader
iterator: initially it distributes the range to split among the tasks
(nTasks). Then it performs adaptive splitting locally on each task. When a
task exhausts its local range, it steals and splits from the range of another
task (victim). The Splitting method on the local range and on the victim range
is Binary (the size of each chunk is the number of unassigned
iterations divided by 2). There are three stealing strategies that can be
selected at compile time with config param methodStealing:

* methodStealing=0: Each task without work tries to steal from its neighbor
     range until it exhausts that range. Then it continues with the next
     neighbor range, and so on until there is not more work. It is the default
     policy.

* methodStealing=1: Each task without work tries to steal once from its
       neighbor range, next from the following neighbor range and so on in a
       round-robin way until there is not more work.

* methodStealing=2: As method 0, but now the splitting in the victi range is
       performed from its tail.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19030 3a8e244f-b0f2-452b-bcba-4c88e055c3ca