Fixed 2 race conditions in kernel 4.
Fixed 2 race conditions in my kernel 4 implementation that were due to
replicating the level set work queues.  Both required a barrier to
insure that globally visible writes were completed before proceeding
to the next iteration.

This never showed up during nightly testing of the RMAT graph, because
the average degree is largish and the paths are shortish and so we got
lucky.  Torus graph testing hit one or both of the race conditions
every time, since the average degree is small and the paths can be
very long.

After a few nights of successful testing, I will promote this to the
top level version.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19032 3a8e244f-b0f2-452b-bcba-4c88e055c3ca