Add 2 feature request futures for a module/sub-module warning message.
The following situation is subtle and it would be nice to have a
warning:

     blah.chpl
     ---------
     module blah {
       ...
     };

The semicolon makes the compiler think that it needs to create a new
top-level module, and thus any symbols in the module will not be
visible unless you do a "use blah.blah".

See r18043 for a similar warning for top-level use statements.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19040 3a8e244f-b0f2-452b-bcba-4c88e055c3ca