Added divceil(m,n) (formerly ceil(m,n)), divfloor(m,n), mod(m,n), mod(x,y).
Also:

* added a test to lock the behavior of the integer operations
(the test ensures it matches the corresponding operations on reals)

* modified exercises/histogram* to use divceil()

* renamed a future test

* modified/added entries to the Standard Modules chapter of the spec

* added (undocumented) isNonnegative(i:integral) to ChapelBase.chpl -
it produces a param result whenever possible,
i.e. when the argument is an unsigned int or a param

Explanations: 

divceil/divfloor compute ceiling/floor of the fraction between two integers.
The naming distinguishes them from ceil/floor - single-argument functions.

mod() is a two-argument operation, so 'div' was not prepended.
The definition was taken from wikipedia's entry on "ceiling function".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19041 3a8e244f-b0f2-452b-bcba-4c88e055c3ca