Move the runtime utilities into a standard module.
To move chpl_rt_utils.chpl into a module (now called ChapelRT.chpl) was almost
trivial.  It was necessary to place the calls to .add and .remove in serial
statements, to prevent deadlock on pthread_mutex_lock().

The rest of this change involves getting the initialization order correct.  To
do this, I added another compiler pass called addInitGuards, which makes each of
the compiler-generated module initialization functions idempotent.  That allows
us to call chpl__init_chpl__Program in main.c before inserting the first task
into the task table.

TODO: It would be better to call just chpl__init_ChapelRT() in main.c, but to do
this, we have to make each module call the initializtion code for all of the
modules it uses.  Then, calling just the init for the module we're using will
suffice.
I didn't remove the fRuntime flag.  But after this checkin, we don't use it
anymore, so it can be removed.
Internal modules are still being treated specially, but need not be.  After
adding an init call for each use to the init function for the module using it,
all we have to do is implicitly use ChapelStandard at the beginning of the main module.

DETAILS:
Moved chpl_rt_util.chpl from runtime/src/common-chpl to modules/internal/ChapelRT.chpl.
Added the new compiler pass addInitGuards:
    For each module initialization function, add and test an initialized?
    predicate to determine whether to run the initialization code.  This makes
    each such function idempotent.
    As a bookkeeping task, add the function chpl__init_preInit() and populate it
    with code that set each of the initialization predicates initially to
    false.  This routine must be (and is) called before any of the module
    initialization functions.
Instead of calling chpl_int_chpl_rt_utils() in main.c, we call
chpl__init_chpl__Program().  This is a sledgehammer to make sure that
chpl__init_ChapelRT() and any modules on which it depends are called before the
first task is inserted in the task table.  After the above TODOs are completed,
we will call just chpl__init_ChapelRT(), and it will "just work".

Factored initTheProgram() out of initPrimitiveTypes(); the latter was becoming a
catch-all.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19044 3a8e244f-b0f2-452b-bcba-4c88e055c3ca