Replace maxThreadsPerLocale with numThreadsPerLocale
This commit replaces the config const 'maxThreadsPerLocale' with
'numThreadsPerLocale' in order to make it more useful for tasking
layers that don't have a varying number of tasks per locale, such as
qthreads and nanox.  For tasking layers that do, such as fifo, it
still serves as a user-specified maximum.  Updated README.tasks (and
generally took a pass over its contents) and README.executing to
reflect this change and document the default behavior for each tasking
layer.

There is still internally a maxThreadsPerLocale-style variable that is
used to store the upper limit on the number of threads to used,
currently as defined by the communication layer (e.g., GASNet's limit
of 255 or 127 threads).  To mark it as internal, I renamed it to
chpl__maxThreadsPerLocale, however.  I also removed the functions in
the tasking layer that let it weigh in on what the maximum should be
because (a) the tasking layers typically did not use these routines,
and (b) if they do have a preference, they can take it into account
directly in chpl_task_init(), rather than passing it to themselves via
a somewhat convoluted path.

To that end, I updated the chpl_task_init() call to take these two
values, as well as an indication of whether the communication layer
was planning on making a polling task, and the pre-existing
callStackSize constant.  Updated all tasking layers and communication
layers to reflect these interface changes.  Our threading layer
interface also changed to take the numThreads/maxThreads values so I
updated those as well.

Updated all tests that referred to maxThreadsPerLocale to refer to
numThreadsPerLocale instead.

Removed primitives that used to be used for these purposes, replacing
them with extern routines.

Renamed chpl_numCoresPerLocale() to chpl_numCoresOnThisLocale() to
reflect the fact that different locales may have different numbers of
processors.  I also moved the caching of this value to avoid calling
into the function too often into the C code for direct calls to it
(such as those we make in the runtime).

Switched instances of _extern in the modules I was touching to extern.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19051 3a8e244f-b0f2-452b-bcba-4c88e055c3ca