Added Map::get_record(); added comments to Map and Vec.
Description:

    MapElem<K,C>* Map<K,C>::get_record(K akey)

This function returns a pointer to the MapElem object ("record") for the key
'akey' if the key exists in the map, or NULL otherwise.

The MapElem object contains both the key and its associated value
(see class MapElem in compiler/include/map.h).

The value can be changed by modifying the field of MapElem.

The key should not be modified because a Map is a closed hashtable -
modifying a key (or setting it to 0, which I think is the "absent" marker)
may result in early termination of probing and not finding the right record.

This function is a trivial (but valuable) wrapper for Vec::set_in.
Feel free to rename the function if you feel strongly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca