Fixed PrgEnv troubles issue
Fixed a problem that has been plaguing our XT PrgEnv testing on
chplbld01 since we moved there due to the fact that unloading one
PrgEnv and loading another seems to leave some cruft around in the
environment.  It's not clear yet whether this is a problem of ours (I
think Brian K. hacked some of the module environment code to get
around other issues) or the modules themselves.  In any event, this
got things working better...

...But not perfectly.  Our compiler turns up what seems to be a bug in
the PGI compiler (which has been breaking our compiler builds ever
since the above fix started getting used -- I have a patch that works
around it but haven't tested yet); and we seem to be having issues on
CCE with -O3 or -O2 on, though I haven't had a chance to get to the
bottom of that one yet.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19095 3a8e244f-b0f2-452b-bcba-4c88e055c3ca