Changes for PGI compatability
When compiling the Chapel compiler with PGI's pgCC, we've been getting
a linktime error due to conflicting definitions of int64_t ('long' vs
'long long' depending on whether stdint.h or cstdlib (or sys/types.h)
is included first.  This commit removes references to sys/types.h and
ensures that cstdlib is the first included in any file that ran into
this error.  We've passed this issue along to PGI as it seems like a 
bug.

Also removed a few extra break statements that generated PGI warnings.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19099 3a8e244f-b0f2-452b-bcba-4c88e055c3ca