Fix for GASNet early termination problem
This commit fixes an infinite loop problem that can sometimes happen
for out GASNet layer if a halt is called early in program
initialization -- in particular, before the polling thread starts up.
This problem can be reproduced by putting a halt() as the first line
of ChapelBase.chpl and running on one locale, in which case the halt
(almost?) always goes into an infinite loop.

The fix is simple, which is to use the bit that already exists to
determine whether or not the polling thread is done to assert that
it's done at the beginning of the program, and to turn that bit off
just before starting the polling thread here.  There's a very subtle
race that could occur here (since the bit isn't flipped atomically
with starting the polling thread), but this puts us on much firmer
ground and fixes the case above.

This should also keep things from going as far off the rails when
GASNet testing goes awry by not having as many runaway processes in
the event that timedexec doesn't kill them (as it seems to
occasionally have failed to do these past few days).

There's one aesthetic change I'd like to make to this, but I'll do
that separately.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19101 3a8e244f-b0f2-452b-bcba-4c88e055c3ca