Reverse sense of the polling flag
With my last commit, the name of the pollingdone flag started to
feel like a double-negative given how it was initialized and
used, so I switched it to pollingRunning in order to reverse
that problem (saved it for a separate commit to support testing
now that the original issue was fixed, and to make the diffs
cleaner).

In doing this renaming, it also revealed that due to my previous
commit, one of the assignments to it when the creation of the
polling task fails was no longer necessary, so I removed that.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19102 3a8e244f-b0f2-452b-bcba-4c88e055c3ca