Improvements to nightly script, including support for qthreads/nanox testing
For awhile now, we've been doing qthreads/nanox testing using a shadow
copy of the nightly script that I had hacked to do things like set the
LD_LIBRARY_PATH and make the appropriate third-party directory.  The
LD_LIBRARY_PATH changes are no longer required thanks to Greg's rpath
additions, so I finally got around to merging the other changes into
the nightly script to avoid keeping the shadow script around any longer.
The changes to nightly.chap02 and nightly.lnx-hilde are to now refer
directly to the nightly script.

The main impact this will have on all testing is that we now embed the
tasking layer name into the unique strings used to define files and
directories for a given testing run, so I believe this means that
tonight's runs will look as though any existing failures we've had
will be "new" (due to a lack of previous results to compare them to).
This is unfortunate, but seems like a bullet to be bitten, and will
give me a good excuse to review the REGRESSIONS file top-to-bottom in
anticipation of the review.

Other changes:

- added a new -hellos flag that tests all hello, world tests rather
  than just the simplest one.  I considered making the -hello flag do
  this, but still like it to serve as a quick end-to-end test.  The
  -hellos flag I find useful because it tests 1- and multi-locale
  runs, task and data parallelism.  A good sanity check for basic
  functionality easily.

- pulled pvm-related stuff




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19113 3a8e244f-b0f2-452b-bcba-4c88e055c3ca