Minor cleanup: decoupling memleaks testing from examples directory
This is a minor change, but in my testing cleanup last week, I noticed
that we were having the --memleaks flag imply the --examples flag
in the nightly script, which feels inappropriate to me.  So moved this
to an explicit throwing of the --examples flag when we invoke the memory
tests.

Also, added a comment about why we only make the spec tests when
--examples is thrown -- originally I thought this was incorrect, now
I think it's correct but confusing (so hopefully the comment will help)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca