Replaced '_stride' with param 'stride' in a few cases; fixed its type.
* I came across a couple of other cases where 'rangeBase._stride' could
be replaced with 'rangeBase.stride' and acted in most cases.

In general, it is correct (I believe) to replace all uses
of '_stride' with 'stride', but in a couple of places it did not appeal
to me as much. (E.g. when copying fields from one record to another;
although that case could be changed as well.)

* I fixed the same bug with the type of 'param stride' as in r19121.

* Added 'inline' or 'param' to ==(rb1,rb2) (and in one case, both(!)
  well one gotta go in that case).

* Removed chpl__byHelp() as it is unused since r19121.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19122 3a8e244f-b0f2-452b-bcba-4c88e055c3ca