Cleaning up third-party Makefiles
Michael Ferguson pointed out that the qthreads Makefile's clean step
was a bit outdated (copied from GASNet from the days when we didn't
do the build in a separate build/ directory), so I updated all the
clean/clobber targets as follows:

  make clean : removes the specific subdirectory of build/ used for
     your current configuration

  make clobber : removes the build/ and install/ directories altogether

Got all autoconf'd third-party packages using consistent schemes,
including bringing tcmalloc/ closer in-line with the other packages
(Greg hadn't known about the ability to build in a side directory and
reviewed the tcmalloc-related changes).

Also cleaned up the top-level third-party Makefile to propagate clean
and clobber commands to all subdirectories.

As with any Makefile change, this is difficult to test, so give a
shout if you run into surprises.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19135 3a8e244f-b0f2-452b-bcba-4c88e055c3ca