Add new primitives for chpl_comm_get and chpl_comm_put.
This commit adds support for the the above primitives.  Use of the
primitives generates calls to the compiler macros CHPL_COMM_GET and
CHPL_COMM_PUT (defined in
$CHPL_HOME/runtime/include/chpl-comm-compiler-macros.h)

The arguments for chpl_comm_get() are:
    - local ref to data
    - locale id to get from (source)
    - ref to data at source
    - length

The arguments for chpl_comm_put() are:
    - local ref to data
    - locale id to put to (dest)
    - ref to data at dest
    - length

The type of the data being transferred is taken from the first
argument (which is used to compute the size of the transfer).

As with all primitives, use is for advance Chapel programmers only and
may be finicky.  See programs in this commit for examples of how to
transfer data between default arrays.

Along with this commit is a small change to the array_get primitive to
be more aware of reference types.  This change was needed for the
examples.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca