Fixing problem with --version test
Fixed the mysterious problem that first Rafa and more recently
Michael ran into with the test of the --version flag.  Turned
out that the logic in the compiler was watertight as expected.
The problem was simply that I ignore < vs. > in diff output and
so the script that generates the .good was broken in that it
didn't ignore a build version of '0' like the compiler does.
Oops.  This ought to fix the problem.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19158 3a8e244f-b0f2-452b-bcba-4c88e055c3ca