Replaced unstable spec test
The use statement example from the spec has been failing
nondeterministically because it relies on the time required to do a
writeln().  I simplified the example by making it rely only on user
code, not on standard modules in order to focus on the use statement
and not on timers.  This should stabilize it.

Also did some wordsmithing of the description of the use statement
while here.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19160 3a8e244f-b0f2-452b-bcba-4c88e055c3ca