Improving mailing list choice for release testing
Just noticed this past week that release testing mail goes to the
chapel_cronmail alias whether it was successful or not.  Changing
that so that only failures are mailed to the chapel_cronmail list.

For that matter, if other people are seeing other success mails every
night, please call them out.  I'm on both lists so don't notice what
goes where.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19169 3a8e244f-b0f2-452b-bcba-4c88e055c3ca