Range modifications: slicing, offset(), isEmpty().
* range.offset() now returns a new range, instead of modifying
an existing range in place. It also now generates an error
when its invocation is illegal.

* Modified isEmpty() so it can have a concise description, which I
recommend to include in the spec:

  "If the represented sequence is defined, reports whether it is empty.
Otherwise an error is reported."

* The rangeBase constructor now checks that its stride argument is 1
for unstridable ranges. That argument is ignored in such cases,
but it seemed like a nice sanity check.

* When slicing r1(r2), the sign of the resulting stride is now that of 'r1'.

  It had been that of 'r2' since slicing was introduced in r11658.
However, there had been no motivation for this choice, neither in the r11658
commit nor in the spec.

  Right now we are leaning towards having 'r1' take priority in determining
the result (the sign of the stride and and the type of the result).
This commit implements the first part of this.

* The above slicing change affected our code as follows:

 - There is a reliance on the stride direction in the dimensional
   distribution. It is too subtle for me to understand, but I was able
   to fix it:
     test/distributions/dm/r.chpl

 - The test that locked in the stride direction in r11658 had to change:
     test/types/range/deitz/test_range_index3.good

 - An test of an unrelated issue
   ("showing a problem with ranges near type boundaries" - r14444)
   also depends on the stride direction, so I changed it.
     test/types/range/diten/rangeSlice.good

* Also changed the slicing implementation to do fewer computations
in the unstridable case or when abs(stride) is the same in both ranges.

* Also have slicing report an error when the result is not expressible
in the expected return type. E.g. (.. by 2 align 0)(.. by 2 align 1).

  TODO: Another case when we should throw such an error is
(..10:uint)(-5..-4).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19185 3a8e244f-b0f2-452b-bcba-4c88e055c3ca